+++ h3ron-graph/src/algorithm/mod.rs
@@ -1,6 +1,7 @@
 pub mod covered_area;
 pub mod differential_shortest_path;
 mod dijkstra;
+pub mod nearest_graph_nodes;
 pub mod path;
 pub mod shortest_path;
 pub mod within_weight_threshold;
@@ -8,5 +9,6 @@ pub mod within_weight_threshold;
 // re-export all algorithm traits
 pub use covered_area::CoveredArea;
 pub use differential_shortest_path::DifferentialShortestPath;
+pub use nearest_graph_nodes::NearestGraphNodes;
 pub use shortest_path::{ShortestPath, ShortestPathManyToMany};
 pub use within_weight_threshold::{WithinWeightThreshold, WithinWeightThresholdMany};